requirements.txt
streamlit
pandas
openpyxl
docx-mailmerge2
num2words
comtypes
libreoffice